# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KB API",
        "description": "HTTP JSON API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/sections": {
            "get": {
                "tags": [
                    "Sections"
                ],
                "summary": "Получение списка разделов",
                "responses": {
                    "200": {
                        "description": "Список разделов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#components/schemas/SectionResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            },
            "post": {
                "tags": [
                    "Sections"
                ],
                "summary": "Создание раздела (admin, moderator)",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SectionCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Созданный раздел",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#components/schemas/SectionWithFieldsResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            }
        },
        "/sections/{section}": {
            "get": {
                "tags": [
                    "Sections"
                ],
                "summary": "Получение раздела с его полями",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "description": "Идентификатор раздела"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Раздел с его значениями",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#components/schemas/SectionWithFieldsResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            },
            "delete": {
                "tags": [
                    "Sections"
                ],
                "summary": "Удаление раздела (admin, moderator)",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "description": "Идентификатор раздела"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удалено"
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            }
        },
        "/sections/{sections}": {
            "patch": {
                "tags": [
                    "Sections"
                ],
                "summary": "Обновление раздела (admin, moderator)",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "description": "Идентификатор раздела"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SectionUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Обновленный раздел",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#components/schemas/SectionWithFieldsResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            }
        },
        "/sections/{section}/materials": {
            "get": {
                "tags": [
                    "Materials"
                ],
                "summary": "Получение постранично разбитого списка материалов секции",
                "description": "> Стоит обратить внимание, что для материала в списке, будут отображаться только те дополнительные поля,\n    для которых установлен флаг `is_present_in_card: true`. Кроме того, хотя в примере и указаны ключи вида `additionalPropN`,\n    на самом деле это всегда будет uuid свойства, для которого представлено значение. К сожалению, спецификация **OpenApi** не позволяет выразить это явным образом",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "description": "Идентификатор раздела материалов"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Страница материалов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/PageOfMaterialResources"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            },
            "post": {
                "tags": [
                    "Materials"
                ],
                "summary": "Создание материала",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "description": "Идентификатор раздела материалов"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MaterialCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Созданный материал",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#components/schemas/MaterialDetailedResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            }
        },
        "/sections/{section}/materials/{material}": {
            "get": {
                "tags": [
                    "Materials"
                ],
                "summary": "Получение материала",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "description": "Идентификатор раздела материалов"
                    },
                    {
                        "name": "material",
                        "in": "path",
                        "description": "Идентификатор материала"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Материал",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#components/schemas/MaterialDetailedResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            },
            "delete": {
                "tags": [
                    "Materials"
                ],
                "summary": "Удаление материала",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "description": "Идентификатор раздела материалов"
                    },
                    {
                        "name": "material",
                        "in": "path",
                        "description": "Идентификатор удаляемого материала"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удалено"
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            },
            "patch": {
                "tags": [
                    "Materials"
                ],
                "summary": "Обновление материала",
                "parameters": [
                    {
                        "name": "section",
                        "in": "path",
                        "description": "Идентификатор раздела материалов"
                    },
                    {
                        "name": "material",
                        "in": "path",
                        "description": "Идентификатор обновляемого материала"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MaterialUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Обновленный материал",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#components/schemas/MaterialDetailedResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MaterialCreateRequest": {
                "required": [
                    "title"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426655440000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Название материала"
                    }
                },
                "type": "object",
                "additionalProperties": {
                    "type": "object",
                    "anyOf": [
                        {
                            "properties": {
                                "id": {
                                    "type": "string",
                                    "example": "123e4567-e89b-12d3-a456-426655440000"
                                }
                            },
                            "type": "object"
                        },
                        {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "123e4567-e89b-12d3-a456-426655440000"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        {
                            "oneOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "number"
                                },
                                {
                                    "type": "boolean"
                                }
                            ]
                        }
                    ]
                }
            },
            "MaterialUpdateRequest": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426655440000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Название материала"
                    }
                },
                "type": "object",
                "additionalProperties": {
                    "type": "object",
                    "anyOf": [
                        {
                            "properties": {
                                "id": {
                                    "type": "string",
                                    "example": "123e4567-e89b-12d3-a456-426655440000"
                                }
                            },
                            "type": "object"
                        },
                        {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "123e4567-e89b-12d3-a456-426655440000"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        {
                            "oneOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "number"
                                },
                                {
                                    "type": "boolean"
                                }
                            ]
                        }
                    ]
                }
            },
            "SectionCreateRequest": {
                "required": [
                    "title",
                    "is_dictionary",
                    "is_navigation"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426655440000"
                    },
                    "title": {
                        "type": "string",
                        "maximum": "255",
                        "minimum": "1",
                        "example": "Имя раздела"
                    },
                    "is_dictionary": {
                        "type": "boolean"
                    },
                    "is_navigation": {
                        "type": "boolean"
                    },
                    "image": {
                        "nullable": true,
                        "type": "string",
                        "maximum": "255",
                        "minimum": "1",
                        "example": "Имя раздела"
                    },
                    "fields": {
                        "type": "array",
                        "items": {
                            "required": [
                                "title",
                                "type",
                                "required",
                                "is_present_in_card",
                                "is_filterable"
                            ],
                            "properties": {
                                "id": {
                                    "type": "string",
                                    "example": "123e4567-e89b-12d3-a456-426655440000"
                                },
                                "title": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "sort_index": {
                                    "type": "integer"
                                },
                                "required": {
                                    "type": "boolean"
                                },
                                "is_present_in_card": {
                                    "type": "boolean"
                                },
                                "is_filterable": {
                                    "type": "boolean"
                                },
                                "type": {
                                    "oneOf": [
                                        {
                                            "$ref": "#components/schemas/TypeString"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeInteger"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeText"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeWiki"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeFloat"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeBoolean"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeEnum"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeFile"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeList"
                                        }
                                    ]
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "SectionUpdateRequest": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426655440000"
                    },
                    "title": {
                        "type": "string",
                        "maximum": "255",
                        "minimum": "1",
                        "example": "Имя раздела"
                    },
                    "is_dictionary": {
                        "type": "boolean"
                    },
                    "is_navigation": {
                        "type": "boolean"
                    },
                    "image": {
                        "nullable": true,
                        "type": "string",
                        "maximum": "255",
                        "minimum": "1",
                        "example": "Имя раздела"
                    },
                    "fields": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string",
                                    "example": "123e4567-e89b-12d3-a456-426655440000"
                                },
                                "title": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "sort_index": {
                                    "type": "integer"
                                },
                                "required": {
                                    "type": "boolean"
                                },
                                "is_present_in_card": {
                                    "type": "boolean"
                                },
                                "type": {
                                    "oneOf": [
                                        {
                                            "$ref": "#components/schemas/TypeString"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeInteger"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeText"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeWiki"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeFloat"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeBoolean"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeEnum"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeFile"
                                        },
                                        {
                                            "$ref": "#components/schemas/TypeList"
                                        }
                                    ]
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Api key",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "apiKey": []
        }
    ]
}